refactor: moved getRawDirectoryMetadataBytes to directory adapter
